Set Elecraft to data mode instead of sideband
Elecraft products offer a data mode that is different from sideband.
Data mode removes ssb processing settings and remembers a different
bandwidth setting. When using data modes, this should be selected.

The previous control code of MD2 was setting the radios to USB. MD6 sets
the radios to data mode (and for good measure we define DATA_R for
reverse sideband although it does not get used).

The MD6 setting is definitely the correct one for the K[x]3 and KX2
products. From reading the K2 manual and programming manual it seems to
be correct for the K2 as well.

I'm not able to build the project. I'm glad to test an APK on the KX3 if
a test build could be made for me.